+++ backend/app/models.py
@@ -57,7 +57,7 @@ class BOMItem(Base):
     parent_id = Column(UUID(as_uuid=True), nullable=False)
     child_type = Column(String(16), nullable=False)
     child_id = Column(UUID(as_uuid=True), nullable=False)
-    quantity = Column(Numeric(10, 3), nullable=False, default=1)
+    quantity = Column(Integer, nullable=False, default=1)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
 
 class OperationLog(Base):
